requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
pymysql==1.1.2
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic[email]==2.11.9
pytest==8.4.2
httpx==0.28.1
flask==3.1.2
flask-cors==6.0.1
mysql-connector-python==9.4.0
python-dotenv==1.1.1
